# docs/swagger.yaml
definitions:
  errs.ValidationError:
    properties:
      message:
        type: string
    type: object
  params.ArticleVersionResponse:
    properties:
      article_id:
        type: integer
      article_version_id:
        type: integer
      body:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      status:
        type: integer
      tag_relationship_score:
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
  params.CreateArticleResponse:
    properties:
      article_id:
        type: integer
      article_version_id:
        type: integer
    type: object
  params.CreateArticleVersionRequest:
    properties:
      body:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  params.CreateArticleVersionResponse:
    properties:
      article_version_id:
        type: integer
    type: object
...
  params.GetArticleDetailResponse:
    properties:
      archived_version:
        $ref: '#/definitions/params.ArticleVersionResponse'
      created_at:
        type: string
      created_by:
        type: string
      drafted_version:
        $ref: '#/definitions/params.ArticleVersionResponse'
      id:
        type: integer
      published_version:
        $ref: '#/definitions/params.ArticleVersionResponse'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  params.UpdateArticleStatusRequest:
    properties:
      status:
        type: integer
    type: object
...
info:
  contact:
    email: mrezaelange@gmail.com
    name: reza
  description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
    Bearer {token}"'
  title: Content Management System API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: Get all articles with optional filters
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: "article_id:asc | article_id:desc |\tarticle_version_id:asc |
          article_version_id:desc |\tcreated_by:asc | created_by:desc |\tupdated_by:asc
          | updated_by:desc |\ttitle:asc | title:desc |\tstatus:asc | status:desc
          |\tversion:asc | version:desc | created_at:asc | created_at:desc | updated_at:asc
          | updated_at:desc | tag_relationship_score:asc | tag_relationship_score:desc"
        in: query
        items:
          type: string
        name: sorts
        type: array
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Status 0 for draft, 1 for published, 2 for archived (comma-separated,
          integer values)
        in: query
        name: status
        type: integer
      - description: Created by (comma-separated, UUID values)
        in: query
        name: created_by
        type: string
      - description: Updated by (comma-separated, UUID values)
        in: query
        name: updated_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/params.ArticleVersionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get all articles with optional filters
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article with the given parameters
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/params.CreateArticleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new article
      tags:
      - articles
  /articles/{articleID}:
    delete:
      consumes:
      - application/json
      description: Delete an article with the given ID
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete an article by ID
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Get article detail with the given ID
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/params.GetArticleDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get article detail by ID
      tags:
      - articles
  /articles/{articleID}/versions:
    get:
      consumes:
      - application/json
      description: Get all versions of an article with the given ID
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/params.ArticleVersionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get all versions of an article by ID
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article version with reference from an article ID
        with the given parameters
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Create Article Version Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.CreateArticleVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/params.CreateArticleVersionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new article version with reference from an article ID
      tags:
      - articles
  /articles/{articleID}/versions/{articleVersionID}:
    get:
      consumes:
      - application/json
      description: Get article version with the given article ID and article version
        ID
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Article Version ID
        in: path
        name: articleVersionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/params.ArticleVersionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get article version by ID and article ID
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article version with reference from an article ID
        and version ID with the given parameters
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Article Version ID
        in: path
        name: articleVersionID
        required: true
        type: integer
      - description: Create Article Version Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.CreateArticleVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/params.CreateArticleVersionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new article version with reference from an article ID and
        version ID
      tags:
      - articles
  /articles/{articleID}/versions/{articleVersionID}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an article version with the given parameters
      parameters:
      - description: fill with Bearer token. The token can be accessed via api /auth/login.
          For accessing draft, published, and archived articles. If authorization
          is not provided default is showing published articles
        in: header
        name: Authorization
        type: string
      - description: Article Version ID
        in: path
        name: articleVersionID
        required: true
        type: integer
      - description: Update Article Status Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/params.UpdateArticleStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update the status of an article version
      tags:
      - articles
...
swagger: "2.0"